+++ user/api.go
@@ -23,6 +23,8 @@ import (
 	log "github.com/sirupsen/logrus"
 
 	"github.com/mdblp/go-common/clients/status"
+	"github.com/mdblp/shoreline/auth0"
+	"github.com/mdblp/shoreline/schema"
 	"github.com/mdblp/shoreline/token"
 	"github.com/mdblp/shoreline/user/middlewares"
 
@@ -63,6 +65,7 @@ type (
 		auditLogger  *log.Logger
 		loginLimiter LoginLimiter
 		provider     rp.RelyingParty
+		auth0Client  *auth0.Auth0Client
 	}
 	Secret struct {
 		Secret string `json:"secret"`
@@ -314,7 +317,7 @@ func NewConfigFromEnv(log *log.Logger) *ApiConfig {
 }
 
 // New create a new shoreline API config
-func New(cfg *ApiConfig, logger *log.Logger, store Storage, auditLogger *log.Logger) *Api {
+func New(cfg *ApiConfig, logger *log.Logger, store Storage, auditLogger *log.Logger, auth0Client *auth0.Auth0Client) *Api {
 
 	provider := createOidcProvider(logger, cfg, strings.Join([]string{cfg.PublicApiURl, "oauth/callback"}, "/"))
 	api := Api{
@@ -323,6 +326,7 @@ func New(cfg *ApiConfig, logger *log.Logger, store Storage, auditLogger *log.Log
 		logger:      logger,
 		auditLogger: auditLogger,
 		provider:    provider,
+		auth0Client: auth0Client,
 	}
 
 	api.loginLimiter.usersInProgress = list.New()
@@ -536,8 +540,8 @@ func (a *Api) CreateUser(res http.ResponseWriter, req *http.Request) {
 func (a *Api) UpdateUser(res http.ResponseWriter, req *http.Request, vars map[string]string) {
 	log := middlewares.GetLogReq(req)
 	log.Info("processing a update user request")
-
 	sessionToken := sanitizeSessionToken(req)
+	var auth0User = &schema.UserUpdate{}
 	if tokenData, err := a.authenticateSessionToken(req.Context(), sessionToken); err != nil {
 		a.sendError(res, http.StatusUnauthorized, STATUS_UNAUTHORIZED, log, err)
 
@@ -601,10 +605,12 @@ func (a *Api) UpdateUser(res http.ResponseWriter, req *http.Request, vars map[st
 			dupCheck := &User{}
 			if updateUserDetails.Username != nil {
 				updatedUser.Username = *updateUserDetails.Username
+				auth0User.Username = updateUserDetails.Username
 				dupCheck.Username = updatedUser.Username
 			}
 			if updateUserDetails.Emails != nil {
 				updatedUser.Emails = updateUserDetails.Emails
+				auth0User.Emails = &updateUserDetails.Emails
 				dupCheck.Emails = updatedUser.Emails
 			}
 
@@ -626,6 +632,7 @@ func (a *Api) UpdateUser(res http.ResponseWriter, req *http.Request, vars map[st
 				a.sendError(res, http.StatusInternalServerError, STATUS_ERR_UPDATING_USR, log, err)
 				return
 			}
+			auth0User.Password = updateUserDetails.Password
 		}
 
 		if updateUserDetails.Roles != nil {
@@ -643,6 +650,17 @@ func (a *Api) UpdateUser(res http.ResponseWriter, req *http.Request, vars map[st
 		if err := a.Store.UpsertUser(req.Context(), updatedUser); err != nil {
 			a.sendError(res, http.StatusInternalServerError, STATUS_ERR_UPDATING_USR, log, err)
 		} else {
+			if a.auth0Client != nil {
+				if usr, err := a.auth0Client.GetUserById(updatedUser.Id); err == nil && usr != nil {
+					if usr.Username == updatedUser.Username {
+						// no need to update with the same value
+						auth0User.Username = nil
+					}
+					if err := a.auth0Client.UpdateUser(updatedUser.Id, auth0User); err != nil {
+						a.logger.Error("Impossible to update user on Auth0: ", err)
+					}
+				}
+			}
 			a.logAudit(req, tokenData, "update request succedeed for username:%s, and is a clinician one{%t}", updatedUser.Username, updatedUser.IsClinic())
 			log.Infof("update request succedeed for username:%s, and is a clinician one{%t}", updatedUser.Username, updatedUser.IsClinic())
 			a.sendUser(res, updatedUser, tokenData.IsServer)
@@ -681,6 +699,17 @@ func (a *Api) GetUserInfo(res http.ResponseWriter, req *http.Request, vars map[s
 			a.sendError(res, http.StatusInternalServerError, STATUS_ERR_FINDING_USR, log, err)
 
 		} else if len(results) == 0 {
+			// check directly in Aut0
+			if a.auth0Client != nil {
+				auth0Usr, err := a.auth0Client.GetUser(user.Username)
+				if err != nil {
+					a.logger.Error("Query to Auth0 failed: ", err)
+				} else if auth0Usr != nil {
+					foundUser := &User{Id: auth0Usr.UserID, Username: auth0Usr.Username, Roles: auth0Usr.Roles, Emails: auth0Usr.Emails}
+					a.sendUser(res, foundUser, tokenData.IsServer)
+					return
+				}
+			}
 			a.sendError(res, http.StatusNotFound, STATUS_USER_NOT_FOUND, log)
 
 		} else if len(results) != 1 {